Rename files to avoid naming conflicts
