+++ frontend/src/App.tsx
@@ -35,7 +35,7 @@ import ModManagementPage from './pages/ModManagementPage';
 import MunicipalDashboardPage from './pages/MunicipalDashboardPage';
 import CheckEmailPage from './pages/CheckEmailPage';
 import { SegmentPage } from './pages/SegmentPage';
-import SubgroupManagementPage from './pages/SubgroupManagementPage';
+import SubGroupCreationPage from './pages/SubGroupCreationPage';
 
 //* Added for SubGroup Management Page
 import SubGroupManagementPage from './pages/SubGroupManagementPage';
@@ -123,8 +123,8 @@ function App() {
                     />
                     
                     <AdminRoute
-                        path={ROUTES.SUBGROUP_MANAGEMENT}
-                        component={SubgroupManagementPage}
+                        path={ROUTES.SUBGROUP_MANAGEMENT_CREATION}
+                        component={SubGroupCreationPage}
                     />
 
                     <CustomRoute
