+++ frontend/src/lib/types/data/subgroupCreation.type.ts
