+++ frontend/src/pages/SubGroupCreationPage.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { useAllSegments } from 'src/hooks/segmentHooks';
 import { UserProfileContext } from '../contexts/UserProfile.Context';
-import SubgroupManagementContent from '../components/content/SubgroupManagementContent';
+import SubgroupManagementContent from '../components/content/SubgroupCreationContent';
 import LoadingSpinner from '../components/ui/LoadingSpinner';
 
 export default function SubgroupManagementPage() {
